feat: infra-resolver added in admin module
